some test file reorganization
